01_pytorch_setup/requirements.txt
# YOLOv11 PyTorch 环境依赖
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
requests>=2.28.0